Move string parsing wirepath code into a separate file.
