+++ wirepath/BUILD.bazel
@@ -4,6 +4,7 @@ go_library(
     name = "go_default_library",
     srcs = [
         "wirepath.go",
+        "wirepath_from_string.go",
         "wirepath_protostrings.go",
         "wirepath_value.go",
     ],
@@ -11,7 +12,6 @@ go_library(
     visibility = ["//visibility:public"],
     deps = [
         "//proto/wirepath:go_default_library",
-        "@com_github_golang_glog//:go_default_library",
         "@org_golang_google_protobuf//proto:go_default_library",
         "@org_golang_google_protobuf//reflect/protoreflect:go_default_library",
     ],
